spit bool and character value
