+++ runtime/interpreter/value.go
@@ -631,391 +631,6 @@ func (VoidValue) ChildStorables() []atree.Storable {
 	return nil
 }
 
-// BoolValue
-
-type BoolValue bool
-
-var _ Value = BoolValue(false)
-var _ atree.Storable = BoolValue(false)
-var _ EquatableValue = BoolValue(false)
-var _ HashableValue = BoolValue(false)
-
-const TrueValue = BoolValue(true)
-const FalseValue = BoolValue(false)
-
-func AsBoolValue(v bool) BoolValue {
-	if v {
-		return TrueValue
-	}
-	return FalseValue
-}
-
-func (BoolValue) isValue() {}
-
-func (v BoolValue) Accept(interpreter *Interpreter, visitor Visitor) {
-	visitor.VisitBoolValue(interpreter, v)
-}
-
-func (BoolValue) Walk(_ *Interpreter, _ func(Value)) {
-	// NO-OP
-}
-
-func (BoolValue) StaticType(interpreter *Interpreter) StaticType {
-	return NewPrimitiveStaticType(interpreter, PrimitiveStaticTypeBool)
-}
-
-func (BoolValue) IsImportable(_ *Interpreter) bool {
-	return sema.BoolType.Importable
-}
-
-func (v BoolValue) Negate(_ *Interpreter) BoolValue {
-	if v == TrueValue {
-		return FalseValue
-	}
-	return TrueValue
-}
-
-func (v BoolValue) Equal(_ *Interpreter, _ LocationRange, other Value) bool {
-	otherBool, ok := other.(BoolValue)
-	if !ok {
-		return false
-	}
-	return bool(v) == bool(otherBool)
-}
-
-func (v BoolValue) Less(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	o, ok := other.(BoolValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-
-	return !v && o
-}
-
-func (v BoolValue) LessEqual(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	o, ok := other.(BoolValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-
-	return !v || o
-}
-
-func (v BoolValue) Greater(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	o, ok := other.(BoolValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-
-	return v && !o
-}
-
-func (v BoolValue) GreaterEqual(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	o, ok := other.(BoolValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-
-	return v || !o
-}
-
-// HashInput returns a byte slice containing:
-// - HashInputTypeBool (1 byte)
-// - 1/0 (1 byte)
-func (v BoolValue) HashInput(_ *Interpreter, _ LocationRange, scratch []byte) []byte {
-	scratch[0] = byte(HashInputTypeBool)
-	if v {
-		scratch[1] = 1
-	} else {
-		scratch[1] = 0
-	}
-	return scratch[:2]
-}
-
-func (v BoolValue) String() string {
-	return format.Bool(bool(v))
-}
-
-func (v BoolValue) RecursiveString(_ SeenReferences) string {
-	return v.String()
-}
-
-func (v BoolValue) MeteredString(memoryGauge common.MemoryGauge, _ SeenReferences) string {
-	if v {
-		common.UseMemory(memoryGauge, common.TrueStringMemoryUsage)
-	} else {
-		common.UseMemory(memoryGauge, common.FalseStringMemoryUsage)
-	}
-
-	return v.String()
-}
-
-func (v BoolValue) ConformsToStaticType(
-	_ *Interpreter,
-	_ LocationRange,
-	_ TypeConformanceResults,
-) bool {
-	return true
-}
-
-func (v BoolValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
-	return v, nil
-}
-
-func (BoolValue) NeedsStoreTo(_ atree.Address) bool {
-	return false
-}
-
-func (BoolValue) IsResourceKinded(_ *Interpreter) bool {
-	return false
-}
-
-func (v BoolValue) Transfer(
-	interpreter *Interpreter,
-	_ LocationRange,
-	_ atree.Address,
-	remove bool,
-	storable atree.Storable,
-	_ map[atree.StorageID]struct{},
-) Value {
-	if remove {
-		interpreter.RemoveReferencedSlab(storable)
-	}
-	return v
-}
-
-func (v BoolValue) Clone(_ *Interpreter) Value {
-	return v
-}
-
-func (BoolValue) DeepRemove(_ *Interpreter) {
-	// NO-OP
-}
-
-func (v BoolValue) ByteSize() uint32 {
-	return 1
-}
-
-func (v BoolValue) StoredValue(_ atree.SlabStorage) (atree.Value, error) {
-	return v, nil
-}
-
-func (BoolValue) ChildStorables() []atree.Storable {
-	return nil
-}
-
-// CharacterValue
-
-// CharacterValue represents a Cadence character, which is a Unicode extended grapheme cluster.
-// Hence, use a Go string to be able to hold multiple Unicode code points (Go runes).
-// It should consist of exactly one grapheme cluster
-type CharacterValue string
-
-func NewUnmeteredCharacterValue(r string) CharacterValue {
-	return CharacterValue(r)
-}
-
-func NewCharacterValue(
-	memoryGauge common.MemoryGauge,
-	memoryUsage common.MemoryUsage,
-	characterConstructor func() string,
-) CharacterValue {
-	common.UseMemory(memoryGauge, memoryUsage)
-
-	character := characterConstructor()
-	return NewUnmeteredCharacterValue(character)
-}
-
-var _ Value = CharacterValue("a")
-var _ atree.Storable = CharacterValue("a")
-var _ EquatableValue = CharacterValue("a")
-var _ ComparableValue = CharacterValue("a")
-var _ HashableValue = CharacterValue("a")
-var _ MemberAccessibleValue = CharacterValue("a")
-
-func (CharacterValue) isValue() {}
-
-func (v CharacterValue) Accept(interpreter *Interpreter, visitor Visitor) {
-	visitor.VisitCharacterValue(interpreter, v)
-}
-
-func (CharacterValue) Walk(_ *Interpreter, _ func(Value)) {
-	// NO-OP
-}
-
-func (CharacterValue) StaticType(interpreter *Interpreter) StaticType {
-	return NewPrimitiveStaticType(interpreter, PrimitiveStaticTypeCharacter)
-}
-
-func (CharacterValue) IsImportable(_ *Interpreter) bool {
-	return sema.CharacterType.Importable
-}
-
-func (v CharacterValue) String() string {
-	return format.String(string(v))
-}
-
-func (v CharacterValue) RecursiveString(_ SeenReferences) string {
-	return v.String()
-}
-
-func (v CharacterValue) MeteredString(memoryGauge common.MemoryGauge, _ SeenReferences) string {
-	l := format.FormattedStringLength(string(v))
-	common.UseMemory(memoryGauge, common.NewRawStringMemoryUsage(l))
-	return v.String()
-}
-
-func (v CharacterValue) NormalForm() string {
-	return norm.NFC.String(string(v))
-}
-
-func (v CharacterValue) Equal(_ *Interpreter, _ LocationRange, other Value) bool {
-	otherChar, ok := other.(CharacterValue)
-	if !ok {
-		return false
-	}
-	return v.NormalForm() == otherChar.NormalForm()
-}
-
-func (v CharacterValue) Less(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	otherChar, ok := other.(CharacterValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-	return v.NormalForm() < otherChar.NormalForm()
-}
-
-func (v CharacterValue) LessEqual(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	otherChar, ok := other.(CharacterValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-	return v.NormalForm() <= otherChar.NormalForm()
-}
-
-func (v CharacterValue) Greater(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	otherChar, ok := other.(CharacterValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-	return v.NormalForm() > otherChar.NormalForm()
-}
-
-func (v CharacterValue) GreaterEqual(_ *Interpreter, other ComparableValue, _ LocationRange) BoolValue {
-	otherChar, ok := other.(CharacterValue)
-	if !ok {
-		panic(errors.NewUnreachableError())
-	}
-	return v.NormalForm() >= otherChar.NormalForm()
-}
-
-func (v CharacterValue) HashInput(_ *Interpreter, _ LocationRange, scratch []byte) []byte {
-	s := []byte(string(v))
-	length := 1 + len(s)
-	var buffer []byte
-	if length <= len(scratch) {
-		buffer = scratch[:length]
-	} else {
-		buffer = make([]byte, length)
-	}
-
-	buffer[0] = byte(HashInputTypeCharacter)
-	copy(buffer[1:], s)
-	return buffer
-}
-
-func (v CharacterValue) ConformsToStaticType(
-	_ *Interpreter,
-	_ LocationRange,
-	_ TypeConformanceResults,
-) bool {
-	return true
-}
-
-func (v CharacterValue) Storable(_ atree.SlabStorage, _ atree.Address, _ uint64) (atree.Storable, error) {
-	return v, nil
-}
-
-func (CharacterValue) NeedsStoreTo(_ atree.Address) bool {
-	return false
-}
-
-func (CharacterValue) IsResourceKinded(_ *Interpreter) bool {
-	return false
-}
-
-func (v CharacterValue) Transfer(
-	interpreter *Interpreter,
-	_ LocationRange,
-	_ atree.Address,
-	remove bool,
-	storable atree.Storable,
-	_ map[atree.StorageID]struct{},
-) Value {
-	if remove {
-		interpreter.RemoveReferencedSlab(storable)
-	}
-	return v
-}
-
-func (v CharacterValue) Clone(_ *Interpreter) Value {
-	return v
-}
-
-func (CharacterValue) DeepRemove(_ *Interpreter) {
-	// NO-OP
-}
-
-func (v CharacterValue) ByteSize() uint32 {
-	return cborTagSize + getBytesCBORSize([]byte(v))
-}
-
-func (v CharacterValue) StoredValue(_ atree.SlabStorage) (atree.Value, error) {
-	return v, nil
-}
-
-func (CharacterValue) ChildStorables() []atree.Storable {
-	return nil
-}
-
-func (v CharacterValue) GetMember(interpreter *Interpreter, _ LocationRange, name string) Value {
-	switch name {
-	case sema.ToStringFunctionName:
-		return NewHostFunctionValue(
-			interpreter,
-			sema.ToStringFunctionType,
-			func(invocation Invocation) Value {
-				interpreter := invocation.Interpreter
-
-				memoryUsage := common.NewStringMemoryUsage(len(v))
-
-				return NewStringValue(
-					interpreter,
-					memoryUsage,
-					func() string {
-						return string(v)
-					},
-				)
-			},
-		)
-
-	case sema.CharacterTypeUtf8FieldName:
-		common.UseMemory(interpreter, common.NewBytesMemoryUsage(len(v)))
-		return ByteSliceToByteArrayValue(interpreter, []byte(v))
-	}
-	return nil
-}
-
-func (CharacterValue) RemoveMember(_ *Interpreter, _ LocationRange, _ string) Value {
-	// Characters have no removable members (fields / functions)
-	panic(errors.NewUnreachableError())
-}
-
-func (CharacterValue) SetMember(_ *Interpreter, _ LocationRange, _ string, _ Value) bool {
-	// Characters have no settable members (fields / functions)
-	panic(errors.NewUnreachableError())
-}
-
 // StringValue
 
 type StringValue struct {
